+++ packages/core/package.json
@@ -1,15 +1,21 @@
 {
   "name": "@opensouls/core",
   "private": false,
-  "version": "0.0.7",
+  "version": "0.0.10",
   "type": "module",
   "description": "The core functionality of the OPEN SOULS soul-engine",
   "main": "dist/index.mjs",
   "types": "dist/types.d.ts",
   "exports": {
-    "import": "./dist/index.mjs",
-    "require": "./dist/index.cjs",
-    "types": "./dist/types.d.ts"
+    ".": {
+      "import": "./dist/index.mjs",
+      "require": "./dist/index.cjs",
+      "types": "./dist/types.d.ts"
+    },
+    "./minimal": {
+      "import": "./dist/minimal.mjs",
+      "types": "./dist/types.d.ts"
+    }
   },
   "scripts": {
     "prepublishOnly": "npm run build",
