+++ packages/core/scripts/build.ts
@@ -54,6 +54,13 @@ await esbuild.build({
   outExtension: { ".js": ".cjs" },
 })
 
+await esbuild.build({
+  ...defaultParams,
+  entryPoints: ["src/minimal.ts"],
+  format: 'esm',
+  outExtension: { ".js": ".mjs" },
+})
+
 await $`cp src/sharedTypes/cortexStep.d.ts dist/`
 
 await $`rm -rf lib temp`
