+++ packages/core/src/index.ts
@@ -5,5 +5,6 @@ export * from './processors/index.js'
 export * from './ReusableStream.js'
 export * from './utils.js'
 export * from './WorkingMemory.js'
+export * from './Memory.js'
 
 export * from "./sharedTypes/index.js"
