+++ packages/core/src/processors/OpenAIProcessor.ts
@@ -7,7 +7,7 @@ import { ChatMessage } from "gpt-tokenizer/GptEncoding";
 import { ZodError, fromZodError } from 'zod-validation-error';
 
 import { registerProcessor } from "./registry.js";
-import { ChatMessageRoleEnum, Memory } from "../WorkingMemory.js";
+import { ChatMessageRoleEnum, Memory } from "../Memory.js";
 import { ChatCompletionMessageParam } from "openai/resources/chat/completions";
 import { ReusableStream } from "../ReusableStream.js";
 import {
